created set_env_var_int for integer values to avoid unnecessary memory allocations
