+++ src/0_core/general_utils_1.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   general_utils_1.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: astein <astein@student.42lisboa.com>       +#+  +:+       +#+        */
+/*   By: anshovah <anshovah@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/25 16:43:17 by anshovah          #+#    #+#             */
-/*   Updated: 2023/11/10 17:06:31 by astein           ###   ########.fr       */
+/*   Updated: 2023/11/10 23:09:47 by anshovah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,11 +33,7 @@ void    reset_cycle(t_mbox *mbox)
 
 void    err_free_and_close_box(t_mbox *mbox, int exit_status)
 {
-    char    *str_exit_status;
-
-    str_exit_status = ft_itoa(exit_status);
-    set_var_value(mbox, "?", str_exit_status);
-    free(str_exit_status);
+    set_var_value_int(mbox, "?", exit_status);
     free_and_close_box_v2(mbox);
 }
