+++ src/0_core/signals.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   signals.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: astein <astein@student.42lisboa.com>       +#+  +:+       +#+        */
+/*   By: anshovah <anshovah@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/25 19:30:56 by anshovah          #+#    #+#             */
-/*   Updated: 2023/11/10 18:35:10 by astein           ###   ########.fr       */
+/*   Updated: 2023/11/10 23:04:05 by anshovah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
